+++ Paramount.Betterclassifieds.Tests/Events/EventTicketFactoryTest.cs
@@ -11,7 +11,7 @@ namespace Paramount.Betterclassifieds.Tests.Events
         public void Create_ReturnsNew_EventTicket()
         {
             var factory = new EventTicketFactory();
-            var result = factory.Create(10, 1, "Adult", 100, isActive: true);
+            var result = factory.Create(10, 1, "Adult", 100, "#blue", isActive: true);
 
             Assert.That(result, Is.TypeOf<EventTicket>());
             Assert.That(result.EventId, Is.EqualTo(1));
@@ -24,6 +24,7 @@ namespace Paramount.Betterclassifieds.Tests.Events
             Assert.That(result.EventTicketReservations.Count, Is.EqualTo(0));
             Assert.That(result.EventBookingTickets, Is.Not.Null);
             Assert.That(result.EventBookingTickets.Count, Is.EqualTo(0));
+            Assert.That(result.ColourCode, Is.EqualTo("#blue"));
 
         }
     }